refactor: rename hub module to executor
